Renamed RequestUtils to MockRequests
